Add modules timeline report tests
