+++ src/main/kotlin/io/github/janbarari/gradle/analytics/metric/modulesbuildheatmap/report/CreateModulesBuildHeatmapReportStage.kt
@@ -20,7 +20,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package io.github.janbarari.gradle.analytics.metric.modulesbuildheatmap.render
+package io.github.janbarari.gradle.analytics.metric.modulesbuildheatmap.report
 
 import io.github.janbarari.gradle.analytics.domain.model.metric.BuildMetric
 import io.github.janbarari.gradle.analytics.domain.model.report.ModuleBuildHeatmap
