+++ src/main/kotlin/io/github/janbarari/gradle/analytics/metric/modulestimeline/report/CreateModulesTimelineReportStage.kt
@@ -20,7 +20,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package io.github.janbarari.gradle.analytics.metric.modulestimeline.render
+package io.github.janbarari.gradle.analytics.metric.modulestimeline.report
 
 import io.github.janbarari.gradle.analytics.domain.model.report.ModulesTimelineReport
 import io.github.janbarari.gradle.analytics.domain.model.report.Report
